Move models package to its own shared module in pkg/models
