+++ go.work
@@ -1,0 +1,8 @@
+go 1.24.4
+
+use (
+    "./apps/api"
+    "./apps/lambdas/upload-processor"
+    "./apps/enclave-service"
+    "./pkg/models"
+)
