+++ pkg/models/go.mod
@@ -1,0 +1,3 @@
+module github.com/QodeSrl/gardbase/pkg/models
+
+go 1.24.4
